+++ NEWS.md
@@ -1,5 +1,6 @@
 # sigminer 1.0.8
 
+- Supported mouse genome 'mm10' ([#241](https://github.com/ShixiangWang/sigminer/issues/241)).
 - Added `read_copynumber_seqz()` to read sequenza result directory.
 - Added new and experimental classification method 'T' for copy number data.
 - Speed up the annotation process in `read_copynumber()`.